Refactor type handling and update car parameter branding.
Ensure consistent float type usage in calculations to avoid type-related issues and improve clarity. Rename `carName` parameter to `brand` for better abstraction and alignment with conventions. These changes enhance code readability and maintainability.

+++ opendbc/car/perodua/carcontroller.py
@@ -227,6 +227,6 @@ class CarController(CarControllerBase):
 
     self.last_steer = apply_steer
     new_actuators = actuators.as_builder()
-    new_actuators.steer = apply_steer / steer_max_interp
+    new_actuators.steer = float(apply_steer / steer_max_interp)
 
     return new_actuators, can_sends
